requirements.txt
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
executing==2.2.1
idna==3.11
ipdb==0.13.13
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
prompt_toolkit==3.0.52
pudb==2025.1.3
pure_eval==0.2.3
Pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
urwid==3.0.3
urwid_readline==0.15.1
wcwidth==0.2.14
